Fix the default integer is 4 bytes bug
As in SQLite3, there's no way to alter the attribute type. The
migration change_phone_number_type_to_string cannot be applied due
to the out of range error as the original phone numbers are in
4-byte integer by default. The following tricks are used to overcome
this issue:
  echo '.dump' | sqlite3 production.sqlite3 > t.dump
Modify the office_phone and cell_phone attribute type in t.dump from
integer to varchar(8), the string format. And then
  cat t.dump | sqlite3 p.sqlite3
Override production.sqlite3:
  cp -f p.sqlite3 production.sqlite3